rename routes/ to pages/
